claude_langchain_demos/pyproject.toml
[project]
name = "claude-langchain-demos"
version = "0.1.0"
description = "LangChain and LangGraph implementations of Claude Quickstarts"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters",
    "python-dotenv",
    "faiss-cpu",
    "fastapi",
    "uvicorn",
    "langchain-experimental",
    "langchain-huggingface",
    "sentence-transformers",
    "defusedxml",
    "beautifulsoup4",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
]
